reorganise tests to match the new layout
